refactor: unify agent settings list
