+++ src/app/renderer/i18n/locales/en.settingsPanel.ts
@@ -77,8 +77,10 @@ export const enSettingsPanel = {
     title: 'Agent',
     defaultAgentLabel: 'Default Agent',
     defaultAgentHelp: 'The default AI provider for new tasks and terminals.',
+    agentListLabel: 'Agents',
     agentProviderOrderLabel: 'Agent Menu Order',
-    agentProviderOrderHelp: 'Controls the order of agents shown in menus.',
+    agentProviderOrderHelp: 'Set the default agent, menu order, and local installation status.',
+    defaultBadge: 'Default',
     moveUp: 'Move up',
     moveDown: 'Move down',
     fullAccessLabel: 'Full Access Mode',
