+++ src/app/renderer/i18n/locales/zh-CN.settingsPanel.ts
@@ -76,8 +76,10 @@ export const zhCNSettingsPanel = {
     title: 'Agent',
     defaultAgentLabel: '默认 Agent',
     defaultAgentHelp: '新任务和终端默认使用的 AI 提供方。',
+    agentListLabel: 'Agents',
     agentProviderOrderLabel: 'Agent 菜单排序',
-    agentProviderOrderHelp: '控制菜单中 Agent 的显示顺序。',
+    agentProviderOrderHelp: '设置默认 Agent、菜单排序和本地安装状态。',
+    defaultBadge: '默认',
     moveUp: '上移',
     moveDown: '下移',
     fullAccessLabel: '完全访问模式',
