+++ src/app/renderer/styles/settings-panel.lists.css
@@ -68,43 +68,94 @@
   color: rgba(232, 103, 120, 0.92);
 }
 
-.settings-provider-card .settings-list-item__left {
-  color: var(--cove-text);
-}
-
-.settings-agent-install-list {
+.settings-agent-list-block {
   display: flex;
   flex-direction: column;
-  gap: 6px;
+  gap: 10px;
   width: 100%;
 }
 
-.settings-agent-install-item {
+.settings-agent-list-block__header {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  min-width: 0;
+}
+
+.settings-agent-list-block__header strong {
+  color: var(--cove-text);
+  font-size: calc(13px * var(--cove-ui-font-scale));
+  font-weight: 600;
+}
+
+.settings-agent-list-block__header span {
+  color: var(--cove-text-muted);
+  font-size: calc(12px * var(--cove-ui-font-scale));
+  line-height: 1.35;
+}
+
+.settings-agent-list-item {
   display: flex;
   flex-direction: column;
   gap: 8px;
 }
 
-.settings-agent-install-row {
+.settings-agent-list-row {
   display: flex;
   align-items: center;
   justify-content: space-between;
   gap: 12px;
   min-height: 48px;
-  padding: 8px 10px;
-  border: 1px solid var(--cove-border-subtle);
-  border-radius: 10px;
-  background: var(--cove-field);
+  padding: 8px 10px 8px 8px;
 }
 
-.settings-agent-install-row__identity {
+.settings-agent-default-choice {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  flex: 1 1 auto;
+  min-width: 0;
+  cursor: pointer;
+}
+
+.settings-agent-default-choice input {
+  position: absolute;
+  inline-size: 1px;
+  block-size: 1px;
+  overflow: hidden;
+  clip: rect(0 0 0 0);
+  clip-path: inset(50%);
+  white-space: nowrap;
+}
+
+.settings-agent-default-choice__visual {
+  width: 14px;
+  height: 14px;
+  flex: 0 0 14px;
+  border: 1px solid var(--cove-border);
+  border-radius: 999px;
+  background: transparent;
+  box-shadow: inset 0 0 0 3px var(--cove-field);
+}
+
+.settings-agent-default-choice input:checked + .settings-agent-default-choice__visual {
+  border-color: var(--cove-accent);
+  background: var(--cove-accent);
+}
+
+.settings-agent-default-choice input:focus-visible + .settings-agent-default-choice__visual {
+  outline: 2px solid rgba(94, 156, 255, 0.85);
+  outline-offset: 2px;
+}
+
+.settings-agent-default-choice__label {
   display: flex;
   align-items: center;
   gap: 10px;
   min-width: 0;
 }
 
-.settings-agent-install-row__name {
+.settings-agent-list-row__name {
   min-width: 0;
   overflow: hidden;
   color: var(--cove-text);
@@ -114,10 +165,16 @@
   white-space: nowrap;
 }
 
-.settings-agent-install-row__icon {
+.settings-agent-list-row__icon {
   width: 22px;
   height: 22px;
   flex-basis: 22px;
+}
+
+.settings-agent-list-row__default {
+  color: var(--cove-text-muted);
+  font-size: calc(11px * var(--cove-ui-font-scale));
+  white-space: nowrap;
 }
 
 .settings-panel button.settings-agent-install__button {
@@ -143,6 +200,10 @@
   padding: 0 10px;
   color: rgba(232, 103, 120, 0.92);
   font-size: calc(11px * var(--cove-ui-font-scale));
+}
+
+.settings-provider-card .settings-list-item__left {
+  color: var(--cove-text);
 }
 
 .settings-panel .settings-list-item__remove.secondary {
