+++ src/contexts/settings/presentation/renderer/settingsPanel/AgentSection.tsx
@@ -1,25 +1,36 @@
 import React from 'react'
 import { ChevronDown, ChevronUp } from 'lucide-react'
+import { AgentProviderIcon } from '@app/renderer/components/AgentProviderIcon'
 import { useTranslation } from '@app/renderer/i18n'
 import { AGENT_PROVIDER_LABEL, type AgentProvider } from '@contexts/settings/domain/agentSettings'
-import { CoveSelect } from '@app/renderer/components/CoveSelect'
+import type { AgentProviderAvailability } from '@shared/contracts/dto'
 
 export function AgentSection(props: {
   defaultProvider: AgentProvider
   agentProviderOrder: AgentProvider[]
   agentFullAccess: boolean
+  availabilityByProvider: Record<string, AgentProviderAvailability>
+  installingProvider: AgentProvider | null
+  installErrorByProvider: Record<string, string>
+  isRefreshingAvailability: boolean
   onChangeDefaultProvider: (provider: AgentProvider) => void
   onChangeAgentProviderOrder: (providers: AgentProvider[]) => void
   onChangeAgentFullAccess: (enabled: boolean) => void
+  onInstallProvider: (provider: AgentProvider) => void
 }): React.JSX.Element {
   const { t } = useTranslation()
   const {
     defaultProvider,
     agentProviderOrder,
     agentFullAccess,
+    availabilityByProvider,
+    installingProvider,
+    installErrorByProvider,
+    isRefreshingAvailability,
     onChangeDefaultProvider,
     onChangeAgentProviderOrder,
     onChangeAgentFullAccess,
+    onInstallProvider,
   } = props
 
   const moveProvider = (fromIndex: number, toIndex: number): void => {
@@ -49,66 +60,111 @@ export function AgentSection(props: {
     <div className="settings-panel__section" id="settings-section-agent">
       <h3 className="settings-panel__section-title">{t('settingsPanel.agent.title')}</h3>
 
-      <div className="settings-panel__row">
-        <div className="settings-panel__row-label">
-          <strong>{t('settingsPanel.agent.defaultAgentLabel')}</strong>
-          <span>{t('settingsPanel.agent.defaultAgentHelp')}</span>
-        </div>
-        <div className="settings-panel__control">
-          <CoveSelect
-            id="settings-default-provider"
-            testId="settings-default-provider"
-            value={defaultProvider}
-            options={agentProviderOrder.map(provider => ({
-              value: provider,
-              label: AGENT_PROVIDER_LABEL[provider],
-            }))}
-            onChange={nextValue => {
-              onChangeDefaultProvider(nextValue as AgentProvider)
-            }}
-          />
-        </div>
-      </div>
-
-      <div className="settings-panel__row">
-        <div className="settings-panel__row-label">
-          <strong>{t('settingsPanel.agent.agentProviderOrderLabel')}</strong>
+      <div className="settings-agent-list-block" id="settings-agent-list">
+        <div className="settings-agent-list-block__header">
+          <strong>{t('settingsPanel.agent.agentListLabel')}</strong>
           <span>{t('settingsPanel.agent.agentProviderOrderHelp')}</span>
         </div>
-        <div className="settings-panel__control settings-panel__control--stack">
-          <div className="settings-list-container">
-            {agentProviderOrder.map((provider, index) => (
-              <div
-                key={provider}
-                className="settings-list-item"
-                data-testid={`settings-agent-order-item-${provider}`}
-              >
-                <div className="settings-list-item__left">{AGENT_PROVIDER_LABEL[provider]}</div>
-                <div className="settings-agent-order__actions">
+
+        <div className="settings-list-container">
+          {agentProviderOrder.map((provider, index) => {
+            const availability = availabilityByProvider[provider]
+            const isInstallingProvider = installingProvider === provider
+            const diagnostics = availability?.diagnostics?.join(' ') ?? ''
+            const installError = installErrorByProvider[provider] ?? ''
+            const isUnavailable = availability?.status === 'unavailable'
+            const isMisconfigured = availability?.status === 'misconfigured'
+            const isBusy = isRefreshingAvailability || Boolean(installingProvider)
+            const installLabel = resolveInstallActionLabel(availability, t, isInstallingProvider)
+            const actionStatus = isInstallingProvider
+              ? 'installing'
+              : (availability?.status ?? 'loading')
+
+            return (
+              <div className="settings-agent-list-item" key={provider}>
+                <div
+                  className="settings-list-item settings-agent-list-row"
+                  data-testid={`settings-agent-order-item-${provider}`}
+                >
+                  <div className="settings-agent-order__actions">
+                    <button
+                      type="button"
+                      className="secondary settings-agent-order__action"
+                      data-testid={`settings-agent-order-move-up-${provider}`}
+                      disabled={index === 0}
+                      aria-label={t('settingsPanel.agent.moveUp')}
+                      onClick={() => moveProvider(index, index - 1)}
+                    >
+                      <ChevronUp className="settings-agent-order__icon" aria-hidden="true" />
+                    </button>
+                    <button
+                      type="button"
+                      className="secondary settings-agent-order__action"
+                      data-testid={`settings-agent-order-move-down-${provider}`}
+                      disabled={index === agentProviderOrder.length - 1}
+                      aria-label={t('settingsPanel.agent.moveDown')}
+                      onClick={() => moveProvider(index, index + 1)}
+                    >
+                      <ChevronDown className="settings-agent-order__icon" aria-hidden="true" />
+                    </button>
+                  </div>
+                  <label className="settings-agent-default-choice">
+                    <input
+                      type="radio"
+                      name="settings-default-provider"
+                      value={provider}
+                      data-testid={`settings-default-provider-${provider}`}
+                      checked={defaultProvider === provider}
+                      onChange={() => onChangeDefaultProvider(provider)}
+                    />
+                    <span className="settings-agent-default-choice__visual" aria-hidden="true" />
+                    <span className="settings-agent-default-choice__label">
+                      <AgentProviderIcon
+                        provider={provider}
+                        className="settings-agent-list-row__icon"
+                      />
+                      <strong className="settings-agent-list-row__name">
+                        {AGENT_PROVIDER_LABEL[provider]}
+                      </strong>
+                      {defaultProvider === provider ? (
+                        <span className="settings-agent-list-row__default">
+                          {t('settingsPanel.agent.defaultBadge')}
+                        </span>
+                      ) : null}
+                    </span>
+                  </label>
                   <button
                     type="button"
-                    className="secondary settings-agent-order__action"
-                    data-testid={`settings-agent-order-move-up-${provider}`}
-                    disabled={index === 0}
-                    aria-label={t('settingsPanel.agent.moveUp')}
-                    onClick={() => moveProvider(index, index - 1)}
+                    className="secondary settings-agent-install__button"
+                    data-status={actionStatus}
+                    data-testid={`settings-agent-executable-install-${provider}`}
+                    aria-label={`${AGENT_PROVIDER_LABEL[provider]} ${installLabel}`}
+                    onClick={() => onInstallProvider(provider)}
+                    disabled={!isUnavailable || isBusy}
                   >
-                    <ChevronUp className="settings-agent-order__icon" aria-hidden="true" />
-                  </button>
-                  <button
-                    type="button"
-                    className="secondary settings-agent-order__action"
-                    data-testid={`settings-agent-order-move-down-${provider}`}
-                    disabled={index === agentProviderOrder.length - 1}
-                    aria-label={t('settingsPanel.agent.moveDown')}
-                    onClick={() => moveProvider(index, index + 1)}
-                  >
-                    <ChevronDown className="settings-agent-order__icon" aria-hidden="true" />
+                    {installLabel}
                   </button>
                 </div>
+
+                {installError.length > 0 ? (
+                  <div
+                    className="settings-agent-install-item__error"
+                    data-testid={`settings-agent-executable-install-error-${provider}`}
+                  >
+                    {installError}
+                  </div>
+                ) : null}
+                {isMisconfigured && diagnostics.length > 0 ? (
+                  <div
+                    className="settings-agent-install-item__error"
+                    data-testid={`settings-agent-executable-diagnostics-${provider}`}
+                  >
+                    {diagnostics}
+                  </div>
+                ) : null}
               </div>
-            ))}
-          </div>
+            )
+          })}
         </div>
       </div>
 
@@ -131,4 +187,28 @@ export function AgentSection(props: {
       </div>
     </div>
   )
+}
+
+function resolveInstallActionLabel(
+  availability: AgentProviderAvailability | null | undefined,
+  t: ReturnType<typeof useTranslation>['t'],
+  isInstalling = false,
+): string {
+  if (isInstalling) {
+    return t('settingsPanel.agentExecutable.status.installing')
+  }
+
+  if (!availability) {
+    return t('common.loading')
+  }
+
+  if (availability.status === 'available') {
+    return t('settingsPanel.agentExecutable.status.available')
+  }
+
+  if (availability.status === 'misconfigured') {
+    return t('settingsPanel.agentExecutable.status.misconfigured')
+  }
+
+  return t('settingsPanel.agentExecutable.install')
 }
